services/service_environment/requirements.txt
kafka-python
pymysql